% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RNACI.r
\docType{package}
\name{RNACI}
\alias{RNACI}
\alias{RNACI-package}
\title{RNACI}
\description{
RNACI
}
\details{
TODO

This package is licensed under the permissive 2-clause BSD license.
}
\author{
Drew Schmidt \email{wrathematics@gmail.com} and Wei-Chen Chen
}
\keyword{package}

